deprecation: delete obsolete binlog test
